+++ packages/react/src/components/Box/styles.ts
@@ -1,7 +1,6 @@
 import { styled } from '../../styles'
 
 export const BoxRoot = styled('div', {
-  backgroundColor: '$background',
   borderRadius: '$2',
 
   padding: '$4',
@@ -9,4 +8,19 @@ export const BoxRoot = styled('div', {
   borderStyle: 'solid',
   borderWidth: 'thin',
   borderColor: '$border',
+
+  variants: {
+    variant: {
+      primary: {
+        background: '$white',
+      },
+      secondary: {
+        backgroundColor: '$background',
+      },
+    },
+  },
+
+  defaultVariants: {
+    variant: 'primary',
+  },
 })